remove Load page component, Search page since their functionality now lives in Search Panel and useHomeFetch hook
