+++ src/App.js
@@ -3,15 +3,7 @@ import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import { ThemeProvider, CSSReset, ColorModeProvider } from '@chakra-ui/core';
 import { customTheme } from './theme';
 import firebase, { FirebaseContext } from './firebase';
-import {
-  Forgot,
-  SearchPage,
-  Home,
-  Login,
-  Movie,
-  UserLists,
-  UserList,
-} from './pages';
+import { Forgot, Home, Login, Movie, UserLists, UserList } from './pages';
 import Layout from './components/Layout';
 import useAuth from './hooks/useAuth';
 
@@ -25,7 +17,6 @@ function App() {
             <Layout>
               <Switch>
                 <Route exact path="/" component={Home} />
-                <Route path="/search" component={SearchPage} />
                 <Route path="/movie/:movieid" component={Movie} />
                 <Route path="/login" component={Login} />
                 <Route path="/lists" component={UserLists} />
