+++ src/pages/index.js
@@ -1,9 +1,8 @@
 import Forgot from './Forgot';
-import SearchPage from './SearchPage';
 import Home from './Home';
 import Login from './Login';
 import Movie from './Movie';
 import UserList from './UserList';
 import UserLists from './UserLists';
 
-export { Forgot, SearchPage, Home, Login, Movie, UserList, UserLists };
+export { Forgot, Home, Login, Movie, UserList, UserLists };
